Move UserCollection persistence handling to LocalStorageManager
