+++ Shared/PostList/PostListView.swift
@@ -106,7 +106,6 @@ struct PostListView: View {
 
     private func reloadFromServer() {
         DispatchQueue.main.async {
-            model.collections.clearUserCollection()
             model.fetchUserCollections()
             model.fetchUserPosts()
         }
